use image as debug container
Currently the image only starts the baremetal command without
any environment variables set.

This commit introduces the sourcing of the ironic.conf
of a pod to which the container can be attached as
debug container. The baremetal cli is then directly
usable without further "manual" setting of environment variables.

The implementation requires the start of a login shell for the
environment to be sourced on startup.

Extended documentation to reflect the use of the
feature.

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>